+++ python/GafferArnoldUI/ArnoldOptionsUI.py
@@ -983,7 +983,6 @@ Gaffer.Metadata.registerNode(
 
 		"options.logFileName.value" : [
 
-			"plugValueWidget:type", "GafferUI.FileSystemPathPlugValueWidget",
 			"path:leaf", True,
 			"fileSystemPath:extensions", "txt log",
 			"fileSystemPath:extensionsLabel", "Show only log files",