+++ python/GafferCycles/CyclesShaderBall.py
@@ -48,7 +48,7 @@ class CyclesShaderBall( GafferScene.ShaderBall ) :
 
 		GafferScene.ShaderBall.__init__( self, name )
 
-		self["environment"] = Gaffer.StringPlug( defaultValue = "${GAFFER_ROOT}/resources/hdri/studio.exr" )
+		self["environment"] = Gaffer.FilePathPlug( defaultValue = "${GAFFER_ROOT}/resources/hdri/studio.exr" )
 
 		# Cycles doesn't support primitives spheres
 		self["__sphere"]["type"].setValue( self["__sphere"].Type.Mesh )